Create Joshbenavente.md
Im seeking the ultimate truth and will stop at nothing to acheive my quest goals.I aim to be trustworthy, painfully honest, and first and foremost a good friend to anyone deserving of my friendship.  Either way I aim to treat everyone with the respect that I deserve, up until they show me that their words are not to be trusted.